@slaren 2024-04-13
Previously tried in https://github.com/ggerganov/llama.cpp/pull/1506