@kunnis 2024-04-14
Replying to both of the previous questions.  I was only able to test this on MSVC, so this changes only targets MSVC, since it was the only one I had proof of the performance difference.  This is a different change then in #1506.  This change only impacts when using F16C with MSVC, vs the #1506 targeted all F16C code.  This change show a pretty clear speed up on multiple processors.

I did look at grouping the F16->F32 conversions, but with this change I'm only showing conversions as 1% of the processor time, and I have other issues I'm working on.