@ggerganov 2024-04-22
> Replying to both of the previous questions. I was only able to test this on MSVC, so this changes only targets MSVC, since it was the only one I had proof of the performance difference. This is a different change then in https://github.com/ggerganov/llama.cpp/pull/1506. This change only impacts when using F16C with MSVC, vs the https://github.com/ggerganov/llama.cpp/pull/1506 targeted all F16C code. This change show a pretty clear speed up on multiple processors.

I'm not sure - is the hypothesis that somehow this helps just on Windows? Why would that be the case?

Would be useful to get some extra reports to see if this improvement is observed for others too