@kunnis 2024-04-22
My guess is this is more of a MSVC or not issue, vs Windows or not.

I'm kind of curious why #1506 didn't work.  In theory it's better not to pollute the cache with a 256k table if intrinsics are available.  The intel docs list all of the opcodes as 0.5 to 1 CPI, so the conversion should be fast.

My best guess is that the code generated by ``_cvtsh_ss(x)`` on non-MSVC is somehow slower than ``_mm_cvtss_f32(_mm_cvtph_ps(_mm_cvtsi32_si128(x)))``.  But I don't understand the details of the FOSS build chain, so I can't get an assembly output to compare to the windows assembly.  _cvtsh_ss() has to wrap multiple opcodes, maybe it's wrapper isn't as fast as the MSVC version?