@slaren 2024-04-22
13900k, cl 19.29.30151:

| model                          |       size | backend    |    threads | test       |       master t/s |          PR t/s | speedup |
| ------------------------------ | ---------: | ---------- | ---------: | ---------- | ---------------: | --------------: | ------: |
| llama 7B Q4_0                  |   3.56 GiB | CPU        |         16 | pp 128     |     33.94 ± 1.52 |    26.26 ± 0.80 | 0.77 |
| llama 7B Q4_0                  |   3.56 GiB | CPU        |         16 | tg 64      |     12.25 ± 0.06 |    12.02 ± 0.04 | 0.98 |
| llama 7B Q4_K - Small          |   3.59 GiB | CPU        |         16 | pp 128     |     40.15 ± 0.26 |    40.47 ± 0.03 | 1.01 |
| llama 7B Q4_K - Small          |   3.59 GiB | CPU        |         16 | tg 64      |     11.80 ± 0.02 |    12.99 ± 0.02 | 1.10 |
| llama 7B F16                   |  12.55 GiB | CPU        |         16 | pp 128     |     15.19 ± 0.12 |    15.10 ± 0.11 | 0.99 |
| llama 7B F16                   |  12.55 GiB | CPU        |         16 | tg 64      |      4.79 ± 0.00 |     4.76 ± 0.01 | 0.99 |

So it looks like there is a significant improvement in some quants, but it is not universal, and Q4_0 has a significant degradation. My guess is that this instruction has a significant latency, and its performance depends entirely on where and how it is used. The best way to handle this may be to change the functions of the specific quants that benefit from this to use this instruction. The benefits are likely to be independent of the compiler, both gcc and MSVC compile this to to the same instruction. IMO this needs more testing.