@kunnis 2024-04-22
@slaren  I'm new to the testing process, how did you generate that chart?  I'll investigate what caused Q4_0 to degrade.