@slaren 2024-04-22
I used `llama-bench` with this command line:
```
llama-bench.exe -m models/7B/ggml-model-q4_0.gguf -m models/7B/ggml-model-Q4_K_S.gguf -m models/7B/ggml-model-f16.gguf -p 128 -n 64
```