@slaren 2024-04-22
It is llama 1 7b that I converted and quantized myself. The specific model shouldn't matter too much, I just used what I had around to run a quick test. llama 2 should produce the same results.