@slaren 2024-04-22
I would suggest doing precisely that. I imagine that all CPUs with AVX2 also support F16C, but I don't know for sure, it would be good to check that first.