@kunnis 2024-04-23
https://github.com/ggerganov/llama.cpp/blob/192090bae47960f0d38d4967abe398a5d190057e/ggml-impl.h#L32   always turns on the F16C flag on windows if you have AVX2.   However, reading this GCC doc it seems like it's possible..  https://gcc.gnu.org/onlinedocs/gcc/x86-Options.html    But the only processor that it lists AVX2 but not F16C is the ‘eden-x4’.  All AMD and Intel chips got F16C before AVX2.

We could change the guard on that block to (\_\_AVX2\_\_ && \_\_F16C\_\_) to cover that edge case, and it'll fall back to the AVX code, which has optional F16C support.  Considering almost all processors that support AVX2 also support F16C, I think keeping the code complexity down would be better. and the only users it'd impact is Via X4.   But that kind of change is something someone with more experience then me needs to chime in on.