@kunnis 2024-04-23
I looked at the manual from intel, https://www.intel.com/content/www/us/en/developer/articles/technical/intel-sdm.html

It says that F16C is "3rd Generation Intel® Core™ processors, Intel® Xeon® processor E3-1200 v2 product family, Intel®
Xeon® processor E5 v2 and E7 v2 families." and AVX2 is "Intel® Xeon® processor E3/E5/E7 v3 product families, 4th Generation Intel® Core™ processor family."

So all Intels that have AVX2 have F16C.

I couldn't find a matching statement about AMD.